Create AbstractIssue
